% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sampling_weights.R
\name{add_sampling_weights}
\alias{add_sampling_weights}
\title{Add column with sampling weights to a data frame}
\usage{
add_sampling_weights(data, variable, population_distribution = NULL)
}
\arguments{
\item{data}{A data frame}

\item{variable}{Character string indicating the variable for whose levels sampling weights should be calculated}

\item{population_distribution}{List (or data frame) specifying the population distribution of the levels; default is \code{NULL}, meaning that all levels are assumed to be represented equally}
}
\value{
A data frame
}
\description{
This function adds a new column \code{sampling_weights} to a data frame. The purpose of sampling weights is to adjust for a mismatch between sample and target population with regard to the distribution of a particular categorical variable. The population distribution can be specified, and sampling weights are then calculated to work as adjustment factors. The default is to assume a balanced population distribution (all levels represented equally), but custom population distributions can be specified.
}
\details{
This function takes as input a data frame, where observations (rows) represent a sample from a population. If the distribution of a categorical variable in the sample does not match the (known or assumed) distribution in the population, the function calculated sampling weights for the observations (rows). These account for the mismatch by up-weighting rows of those level that are underrepresented in the sample (relative to the population) and down-weighting rows belonging to levels that are overrepresented in the sample (relative to the population). If no population distribution is specified, all levels are assumed to be represented equally in the target population. Sampling weights are calculated on the basis of (i) the observed distribution of the variable in the sample, and (ii) the population distribution. For instance, if a specific subgroup (i.e. level) has a share of 10\% in the sample, compared to 20\% in the population, the sampling weight is 2.0 (20\% divided by 10\%). Sampling weights above 1 indicate up-weighting, sampling weights below 1 indicate down-weighting. The function prints out information on the sample and population distribution and the resulting weights.
}
\examples{
add_sampling_weights(
  data = ice_metadata,
  variable = "mode")
}
\author{
Lukas Soenning
}
