% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_DMB.R
\name{disp_DMB}
\alias{disp_DMB}
\title{Calculate the dispersion measure \emph{D~MB~}}
\usage{
disp_DMB(
  subfreq,
  partsize,
  directionality = "conventional",
  conf_int = FALSE,
  conf_level = 0.95,
  digits = NULL,
  verbose = TRUE,
  print_score = TRUE,
  suppress_warning = FALSE
)
}
\arguments{
\item{subfreq}{A numeric vector of subfrequencies, i.e. the number of occurrences of the item in each corpus part}

\item{partsize}{A numeric vector specifying the size of the corpus parts}

\item{directionality}{Character string indicating the directionality of scaling. See details below. Possible values are \code{"conventional"} (default) and \code{"gries"}}

\item{conf_int}{Logical. Whether a (profile likelihood) confidence interval should be computed; default: \code{FALSE}}

\item{conf_level}{Scalar giving the confidence level; default \code{0.95} for a 95\% CI}

\item{digits}{Rounding: Integer value specifying the number of decimal places to retain (default: no rounding)}

\item{verbose}{Logical. Whether additional information (on directionality, formulas, frequency adjustment) should be printed; default is \code{TRUE}}

\item{print_score}{Logical. Whether the dispersion score should be printed to the console; default is \code{TRUE}}

\item{suppress_warning}{Logical. Whether warning messages should be suppressed; default is \code{FALSE}}
}
\value{
A vector of numeric values
}
\description{
This function calculates \emph{D~MB~}, a generalized version of the Poisson-based dispersion measure \emph{MB} proposed by Nelson (2025). It allows the user to choose the directionality of scaling, i.e. whether higher values denote a more even or a less even distribution, and it returns confidence intervals for \emph{D~MB~}.
}
\details{
The function calculates the dispersion measure \emph{D~MB~} based on a set of subfrequencies (number of occurrences of the item in each corpus part) and a matching set of part sizes (the size of the corpus parts, i.e. number of word tokens). \emph{D~MB~} can be considered a generalization of the method proposed by Nelson (2025). In contrast to the original measure, \emph{MB}, \emph{D~MB~} works with a pre-determined set of corpus parts, which may also differ in size. To provide this additional flexibility, \emph{D~MB~} is constructed based on a Poisson regression model that considers the corpus parts as observations and allows them to differ in length through its incorporation of an offset parameter.
\itemize{
\item Directionality: \emph{D~MB~} ranges from 0 to 1. The conventional scaling of dispersion measures (see Juilland & Chang-Rodriguez 1964; Carroll 1970; Rosengren 1971) assigns higher values to more even/dispersed/balanced distributions of subfrequencies across corpus parts. This is the default. Gries (2008) uses the reverse scaling, with higher values denoting a more uneven/bursty/concentrated distribution; use \code{directionality = "gries"} to choose this option.
}
}
\examples{
disp_DMB(
  subfreq = c(0,0,1,2,5), 
  partsize = rep(1000, 5),
  directionality = "conventional",
  conf_int = TRUE)
}
\references{
Carroll, John B. 1970. An alternative to Juilland’s usage coefficient for lexical frequencies and a proposal for a standard frequency index. \emph{Computer Studies in the Humanities and Verbal Behaviour} 3(2). 61--65. \doi{doi:10.1002/j.2333-8504.1970.tb00778.x}

Gries, Stefan Th. 2008. Dispersions and adjusted frequencies in corpora. \emph{International Journal of Corpus Linguistics} 13(4). 403--437. \doi{doi:10.1075/ijcl.13.4.02gri}

Juilland, Alphonse G. & Eugenio Chang-Rodríguez. 1964. \emph{Frequency dictionary of Spanish words.} The Hague: Mouton de Gruyter. \doi{doi:10.1515/9783112415467}

Nelson, Robset N. Jr. 2025. Groundhog Day is not a good model for corpus dispersion. Journal of Quantitative Linguistics 32(2). 103--127. \doi{10.1080/09296174.2024.2423415}

Rosengren, Inger. 1971. The quantitative concept of language and its relation to the structure of frequency dictionaries. \emph{Études de linguistique appliquée (Nouvelle Série)} 1. 103--127.
}
\author{
Lukas Soenning
}
