% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_DP.R
\name{disp_DP_bw}
\alias{disp_DP_bw}
\title{Bootstrap (and weight) Gries's \emph{deviation of proportions}}
\usage{
disp_DP_bw(
  subfreq,
  partsize,
  n_replicates = 500,
  boot_ci = FALSE,
  conf_level = 0.95,
  partweight = NULL,
  directionality = "conventional",
  formula = "egbert_etal_2020",
  freq_adjust = FALSE,
  freq_adjust_method = "even",
  unit_interval = TRUE,
  digits = NULL,
  verbose = TRUE,
  print_score = TRUE,
  suppress_warning = FALSE
)
}
\arguments{
\item{subfreq}{A numeric vector of subfrequencies, i.e. the number of occurrences of the item in each corpus part}

\item{partsize}{A numeric vector specifying the size of the corpus parts}

\item{n_replicates}{Integer value specifying the number of bootstrap replicates to draw; default: \code{500}}

\item{boot_ci}{Logical. Whether a percentile bootstrap confidence interval should be computed; default: \code{FALSE}}

\item{conf_level}{Scalar giving the confidence level; default \code{0.95} for a 95\% percentile CI}

\item{partweight}{A numeric vector specifying the weights of the corpus parts; if not specified, function returns unweighted estimate}

\item{directionality}{Character string indicating the directionality of scaling. See details below. Possible values are \code{"conventional"} (default) and \code{"gries"}}

\item{formula}{Character string indicating which formula to use for the calculation of DP. See details below. Possible values are \code{"egbert_etal_2020"} (default), \code{"gries_2008"}, \code{"lijffit_gries_2012"}.}

\item{freq_adjust}{Logical. Whether dispersion score should be adjusted for frequency (i.e. whether frequency should be 'partialed out'); default is \code{FALSE}}

\item{freq_adjust_method}{Character string indicating which method to use for devising dispersion extremes. See details below. Possible values are \code{"even"} (default) and \code{"pervasive"}}

\item{unit_interval}{Logical. Whether frequency-adjusted scores that exceed the limits of the unit interval should be replaced by 0 and 1; default is \code{TRUE}}

\item{digits}{Rounding: Integer value specifying the number of decimal places to retain (default: no rounding)}

\item{verbose}{Logical. Whether additional information (on directionality, formulas, frequency adjustment) should be printed; default is \code{TRUE}}

\item{print_score}{Logical. Whether the dispersion score should be printed to the console; default is \code{TRUE}}

\item{suppress_warning}{Logical. Whether warning messages should be suppressed; default is \code{FALSE}}
}
\description{
This function offers facilities for bootstrapping and weighting Gries's dispersion measure DP (deviation of proportions). In addition to the full functionality offered by the function \code{disp_DP()}, it can be used to obtain weighted dispersion scores as well as bootstrap confidence intervals.
}
\details{
This function calculates weighted dispersion measures and bootstrap confidence intervals.
}
\examples{
disp_DP_bw(
  subfreq = biber150_ice_gb[3,], 
  partsize = biber150_ice_gb[1,], 
  digits = 2,
  freq_adjust = TRUE,
  directionality = "conventional",
  formula = "gries_2008")

}
\seealso{
\code{\link[=disp_DP]{disp_DP()}} for finer control over the calculation of DP
}
\author{
Lukas Soenning
}
