% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_DP.R
\name{disp_DP_sboot}
\alias{disp_DP_sboot}
\title{Stratified bootstrapping for Gries's \emph{deviation of proportions}}
\usage{
disp_DP_sboot(
  text_freq,
  text_size,
  corpus_parts = NULL,
  n_boot = 500,
  boot_ci = FALSE,
  conf_level = 0.95,
  return_distribution = FALSE,
  partweight = NULL,
  directionality = "conventional",
  formula = "egbert_etal_2020",
  freq_adjust = FALSE,
  freq_adjust_method = "even",
  unit_interval = TRUE,
  digits = NULL,
  verbose = TRUE,
  print_score = TRUE,
  suppress_warning = FALSE
)
}
\arguments{
\item{text_freq}{Integer value giving the frequency of the item in the text}

\item{text_size}{Interger value giving the size (or length) of the text}

\item{corpus_parts}{The corpus parts that form the basis of dispersion analysis; must be higher-level categories, above the text files}

\item{n_boot}{Integer value specifying the number of bootstrap samples to draw; default: \code{500}}

\item{boot_ci}{Logical. Whether a percentile bootstrap confidence interval should be computed; default: \code{FALSE}}

\item{conf_level}{Scalar giving the confidence level; default \code{0.95} for a 95\% percentile CI}

\item{return_distribution}{Logical. Whether the function should return a vector of all \code{n_boot} bootstrap statistics instead of a summary measure}

\item{partweight}{A numeric vector specifying the weights of the corpus parts; if not specified, function returns unweighted estimate}

\item{directionality}{Character string indicating the directionality of scaling. See details below. Possible values are \code{"conventional"} (default) and \code{"gries"}}

\item{formula}{Character string indicating which formula to use for the calculation of DP. See details below. Possible values are \code{"egbert_etal_2020"} (default), \code{"gries_2008"}, \code{"lijffit_gries_2012"}.}

\item{freq_adjust}{Logical. Whether dispersion score should be adjusted for frequency (i.e. whether frequency should be 'partialed out'); default is \code{FALSE}}

\item{freq_adjust_method}{Character string indicating which method to use for devising dispersion extremes. See details below. Possible values are \code{"even"} (default) and \code{"pervasive"}}

\item{unit_interval}{Logical. Whether frequency-adjusted scores that exceed the limits of the unit interval should be replaced by 0 and 1; default is \code{TRUE}}

\item{digits}{Rounding: Integer value specifying the number of decimal places to retain (default: no rounding)}

\item{verbose}{Logical. Whether additional information (on directionality, formulas, frequency adjustment) should be printed; default is \code{TRUE}}

\item{print_score}{Logical. Whether the dispersion score should be printed to the console; default is \code{TRUE}}

\item{suppress_warning}{Logical. Whether warning messages should be suppressed; default is \code{FALSE}}
}
\description{
This function implements stratified bootstrapping (and weighting) for Gries's dispersion measure DP (deviation of proportions). In addition to the full functionality offered by the function \code{disp_DP()}, it can be used to obtain weighted dispersion scores as well as bootstrap confidence intervals.
}
\details{
This function performs stratified bootstrapping on dispersion measures. Stratified bootstrapping is used when the corpus parts represent text categories (e.g. genres, registers) that in turn consists of texts or text files. Since the resampling scheme implemented in bootstrapping should be as closely aligned with the data layout (and data-generation procedure) as closely as possible, resampling should not take place at the level of the text categories. Instead, it is the sampling units in corpus compilation -- texts, text files, or speakers -- that should be resampled. Stratified bootstrapping therefore respects the structure of the corpus and data.
}
\examples{
disp_DP_sboot(
  text_freq = biber150_brown[87,], 
  text_size = biber150_brown[1,], 
  corpus_parts = as.character(metadata_brown$genre),
  digits = 2,
  freq_adjust = TRUE,
  directionality = "conventional",
  formula = "gries_2008")

}
\seealso{
\code{\link[=disp_DP]{disp_DP()}} for finer control over the calculation of DP
}
\author{
Lukas Soenning
}
