% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_analysis_helper_functions.R
\name{find_min_disp}
\alias{find_min_disp}
\title{Find the pessimum distribution (dispersion minimum) for a vector of subfrequencies and part sizes}
\usage{
find_min_disp(subfreq, partsize)
}
\arguments{
\item{subfreq}{A numeric vector of subfrequencies, i.e. the number of occurrences of the item in each corpus part}

\item{partsize}{A numeric vector with the size of the corpus parts}
}
\value{
An integer vector of the same length as the number of corpus parts.
}
\description{
This function returns the (hypothetical) distribution of subfrequencies that represents the smallest possible level of dispersion for the item across the corpus parts. It requires a vector of subfrequencies and a vector of corpus part sizes. This distribution is required for the min-max transformation proposed by Gries (2024: 196-208) to obtain frequency-adjusted dispersion scores.
}
\details{
This function creates a hypothetical distribution of the total number of occurrences of the item (i.e. the sum of the subfrequencies) across the corpus parts. To obtain the highest possible level of concentration (i.e. minimum dispersion), the occurrences are allocated to the smallest corpus parts. The function starts by filling in the smallest corpus part. If the total number of occurrences of the items is greater than the size of the smallest corpus part, it then continues with the second smallest part and so on. The function returns a vector of length \eqn{k} (where \eqn{k} is the number of corpus parts), whose order matches that of the vector of corpus parts provided to the argument \code{partsize}.
}
\examples{
\dontrun{
# not run
find_min_disp(
  subfreq = c(0,0,1,2,5), 
  partsize = rep(1000, 5))
}

}
\author{
Lukas Soenning
}
