% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion_analysis_helper_functions.R
\name{find_min_disp_tdm}
\alias{find_min_disp_tdm}
\title{Find the pessimum distribution (dispersion minimum) for each item in a term-document matrix}
\usage{
find_min_disp_tdm(tdm, row_partsize = "first_row")
}
\arguments{
\item{tdm}{A term-document matrix, where rows represent items and columns corpus parts; must also contain a row giving the size of the corpus parts (first or last row in the term-document matrix)}

\item{row_partsize}{Character string indicating which row in the term-document matrix contains the size of the corpus parts. Possible values are \code{'first_row'}, \code{'last_row'} (default)}
}
\value{
A matrix of integers
}
\description{
This function takes as input a term-document matrix and returns, for each item (i.e. row), the (hypothetical) distribution of subfrequencies that represents the smallest possible level of dispersion for the item across the corpus parts. This distribution is required for the min-max transformation proposed by Gries (2024: 196-208) to obtain frequency-adjusted dispersion scores.
}
\details{
This function takes as input a term-document matrix and creates, for each item in the matrix, a hypothetical distribution of the total number of occurrences of the item (i.e. the sum of the subfrequencies) across the corpus parts. To obtain the highest possible level of concentration (i.e. minimum dispersion) for each item, its occurrences are allocated to the smallest corpus parts. The function starts by filling in the smallest part. If the total number of occurrences of the item is greater than the size of this part, it continues with the second smallest corpus part, and so on. The function returns a matrix of integers which has the same structure as the term-document matrix supplied to the function.
}
\examples{
\dontrun{
# not run
find_min_disp(
  subfreq = matrix(
    c(0,0,1,2,5,
      0,1,2,4,4), 
      nrow = 2, 
      byrow = TRUE),
  partsize = c(100, 100, 100, 500, 1000),
  freq_adjust_method = "pervasive")
}

}
\author{
Lukas Soenning
}
