% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_folded_power.R
\name{fpower_trans}
\alias{fpower_trans}
\title{Tukey's folded power transformation}
\usage{
fpower_trans(lambda = 0)
}
\arguments{
\item{lambda}{Numeric value of the applied power transformation, which can range between 0 (limiting case: logit transformation) and 1 (no transformation)}
}
\value{
A numeric vector
}
\description{
Tukey's folded power transformation
}
\examples{
plot(fpower_trans(lambda = .5), xlim = c(0, 1))
}
