% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_ice.R
\docType{data}
\name{metadata_ice}
\alias{metadata_ice}
\title{Text metadata for ICE corpora}
\format{
\subsection{\code{metadata_ice}}{

A data frame with 500 rows and 6 columns:

\describe{
\item{text_file}{Standardized name of the text file (e.g. "s1a-001", "w1b-008", "w2d-018")}
\item{mode}{Mode of production ("spoken" vs. "written")}
\item{text_category}{4 higher-level text categories ("dialogues", "monologues", "non-printed", "printed"); ordered factor}
\item{macro_genre}{12 macro genres (e.g. "private_dialogues", "student_writing", "reportage"); ordered factor}
\item{genre}{32 genres (e.g. "phonecalls", "unscripted_speeches", "novels_short_stories"); ordered factor}
\item{genre_short}{Short label for the genre (see Schützler 2023: 228); ordered factor}
\item{word_count}{The length of the text file, expressed as the number of (word and nonword) tokens}
}
}
}
\source{
\url{https://www.ice-corpora.uzh.ch/en/design.html}

Greenbaum, Sidney. 1996. Introducing ICE. In Sidney Greenbaum (ed.), \emph{Comparing English worldwide: The International Corpus of English}, 3--12. Oxford: Clarendon Press.

Schützler, Ole. 2023. \emph{Concessive constructions in varieties of English}. Berlin: Language Science Press. \doi{doi:10.5281/zenodo.8375010}
}
\usage{
metadata_ice
}
\description{
This dataset provides metadata for the text files in the ICE family of corpora. It maps standardized file names to various textual categories such as mode of production, macro genre and genre. Text categories, macro genres and genres are ordered based on the sampling frame informing the design of the ICE family of corpora (see https://www.ice-corpora.uzh.ch/en/design.html).
}
\keyword{datasets}
