% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_folded_power.R
\name{scale_x_fpower}
\alias{scale_x_fpower}
\title{Position scales for Tukey's folded power transformation}
\usage{
scale_x_fpower(lambda = 0, breaks = NULL, labels = NULL, n_breaks = 6, ...)
}
\arguments{
\item{lambda}{Numeric value of the applied power transformation}

\item{breaks}{Numeric values indicating where the tick marks should be placed}

\item{labels}{Character vector giving the labels that should be drawn at the tick marks}

\item{n_breaks}{Integer specifying the number of tick marks to draw}

\item{...}{Other argument passed on to \verb{scale_(x|y)_continuous()}}
}
\value{
The ggplot2 function \verb{scale_(x|y)_continuous()} with the appropriate transformation
}
\description{
Position scales for Tukey's folded power transformation
}
\details{
This function was written with the help of ChatGPT (version GPT-5.1; OpenAI 2025)
}
\examples{
if (require("ggplot2")) {
  ggplot(
    data = data.frame(
      dispersion = seq(0, 1, .01)),
    aes(x = dispersion)) +
    geom_dotplot() +
    scale_x_fpower(lambda = .5)
}
}
\references{
OpenAI. (2025). ChatGPT (GPT-5.1) Large language model. https://chat.openai.com
}
