% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp_DP.R
\name{disp_DP}
\alias{disp_DP}
\title{Calculate Gries's DP (deviation of proportions)}
\usage{
disp_DP(
  subfreq,
  partsize,
  formula = "egbert_etal_2020",
  directionality = "conventional",
  verbose = TRUE
)
}
\arguments{
\item{subfreq}{A numeric vector of subfrequencies, i.e. the number of occurrences of the item in each corpus part}

\item{partsize}{A numeric vector with the size of the corpus parts}

\item{formula}{Character string indicating which formula to use for the calculation of DP. See details below. Possible values are \code{'egbert_etal_2020'} (default), \code{'gries_2008'}, \code{'lijffit_gries_2012'}}

\item{directionality}{Character string indicating the directionality of scaling. See details below. Possible values are \code{'conventional'} (default), \code{'gries'}}

\item{verbose}{Logical. Whether additional information on scaling and formulas should be printed, default is TRUE}
}
\value{
Returns a numeric value, the dispersion score.
}
\description{
This function calculates Gries's dispersion measure DP (deviation of proportions). It offers three different formulas and allows the user to choose the directionality of scaling, i.e. whether higher values denote a more even or a less even distribution.
}
\details{
The function calculates the dispersion measure DP based on a set of subfrequencies (number of occurrences of the item in each corpus part) and a matching set of part sizes (the size of the corpus parts, i.e. number of word tokens).
\itemize{
\item Directionality: DP ranges from 0 to 1. The \code{conventional} scaling of dispersion measures (see Juilland & Chang-Rodriguez 1964; Carroll 1970; Rosengren 1971) assigns higher values to more even/dispersed/balanced distributions of subfrequencies across corpus parts. Gries (2008) uses the reverse scaling, with higher values denoting a more uneven/bursty/concentrated distribution; this is implemented by the value \code{gries}.
\item Formula: Irrespective of the directionality of scaling, four formulas for DP exist in the literature (see below for details). This is because the original version proposed by Gries (2008: 415), which is commonly denoted as \eqn{DP} (and here referenced by the value \code{gries_2008}) does not always reach its theoretical limits of 0 and 1. For this reason, modifications have been suggested, starting with Gries (2008: 419) himself, who referred to this version as \eqn{DPnorm}. This version is not implemented in the current package, because Lijffit & Gries (2012) updated \eqn{DPnorm} to ensure that it also works as intended when corpus parts differ in size; this version is represented by the value \code{lijffit_gries_2012} and often denoted using subscript notation \eqn{DP_{norm}} . Finally, Egbert et al. (2020: 99) suggest a further modification to ensure proper behavior in settings where the item occurs in only one corpus part. They label this version \eqn{D_P}. In the current function, it is the default and represented by the value \code{egbert_etal_2020}.
}

In the formulas given below, the following notation is used:
\itemize{
\item \eqn{k} the number of corpus parts
\item \eqn{t_i} a proportional quantity; the subfrequency in part \eqn{i} divided by the total number of occurrences of the item in the corpus (i.e. the sum of all subfrequencies)
\item \eqn{w_i} a proportional quantity; the size of corpus part \eqn{i} divided by the size of the corpus (i.e. the sum of a part sizes)
}

The value \code{gries_2008} implements the original version proposed by Gries (2008: 415). Note that while the following formula represents \code{gries} scaling (0 = even, 1 = uneven), in the current function the directionality is controlled separately using the argument \code{directionality}.

  \eqn{\frac{\sum_i^k |t_i - w_i|}{2}} (Gries 2008)

The value \code{lijffit_gries_2012} implements the modified version described by Lijffit & Gries (2012). Again, the following formula represents \code{gries} scaling (0 = even, 1 = uneven), but the directionality is handled separately in the current function. The notation \eqn{min\{w_i\}} refers to the \eqn{w_i} value of the smallest corpus part.

  \eqn{\frac{\sum_i^k |t_i - w_i|}{2} \times \frac{1}{1 - min\{w_i\}}} (Lijffijt & Gries 2012)

The value \code{egbert_etal_2020} (default) selects the modification suggested by Egbert et al. (2020: 99). The following formula represents \code{conventional} scaling (0 = uneven, 1 = even). The notation \eqn{min\{w_i: t_i > 0\}} refers to the \eqn{w_i} value among those corpus parts that include at least one occurrence of the item.

  \eqn{1 - \frac{\sum_i^k |t_i - w_i|}{2} \times \frac{1}{1 - min\{w_i: t_i > 0\}}} (Egbert et al. 2020)
}
\examples{
\dontrun{
# not run
disp_DP(
  subfreq = c(0,0,1,2,5), 
  partsize = rep(1000, 5),
  formula = "gries_2008",
  directionality = "conventional")
}

}
\references{
Carroll, John B. 1970. An alternative to Juilland’s usage coefficient for lexical frequencies and a proposal for a standard frequency index. \emph{Computer Studies in the Humanities and Verbal Behaviour} 3(2). 61–65. \url{https://doi.org/10.1002/j.2333-8504.1970.tb00778.x}

Egbert, Jesse, Brent Burch & Douglas Biber. 2020. Lexical dispersion and corpus design. \emph{International Journal of Corpus Linguistics} 25(1). 89–115. \url{https://doi.org/10.1075/ijcl.18010.egb}

Gries, Stefan Th. 2008. Dispersions and adjusted frequencies in corpora. \emph{International Journal of Corpus Linguistics} 13(4). 403–437. \url{https://doi.org/10.1075/ijcl.13.4.02gri}

Juilland, Alphonse G. & Eugenio Chang-Rodríguez. 1964. \emph{Frequency dictionary of Spanish words.} The Hague: Mouton de Gruyter. \url{https://doi.org/10.1515/9783112415467}

Lijffijt, Jefrey & Stefan Th. Gries. 2012. Correction to Stefan Th. Gries’ ‘Dispersions and adjusted frequencies in corpora’. \emph{International Journal of Corpus Linguistics} 17(1). 147–149. \url{https://doi.org/10.1075/ijcl.17.1.08lij}

Rosengren, Inger. 1971. The quantitative concept of language and its relation to the structure of frequency dictionaries. \emph{Études de linguistique appliquée (Nouvelle Série)} 1. 103–127.
}
\author{
Lukas Soenning
}
