% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_folded_power.R
\name{invfpower}
\alias{invfpower}
\title{Back-transform folded-power-transformed scores to the unit interval [0,1]}
\usage{
invfpower(y, lambda = 1, scaling = "plus_minus_1")
}
\arguments{
\item{y}{A numeric vector of folded-power-transformed scores}

\item{lambda}{Numeric value of the applied power transformation, which can range between 0 (limiting case: logit transformation) and 1 (no transformation)}

\item{scaling}{Character string indicating whether scores were re-expressed to the [-1, 1] interval (\code{plus_minus_1}) or not (\code{free})}
}
\value{
A numeric vector
}
\description{
This function takes as input a vector of transformed scores, i.e. values that were originally in the unit interval [0, 1] but which were re-expressed using Tukey's folded power transformation. It allows back-transformation of two versions of folded powers: Those that are mapped to the [-1, +1] interval and those that aren't.
}
\details{
This function was written with the help of ChatGPT (version GPT-5.1; OpenAI 2025)
}
\examples{
invfpower(
  seq(-1, 1, .1),
  lambda = .14,
  scaling = "plus_minus_1")
}
\references{
OpenAI. (2025). ChatGPT (GPT-5.1) Large language model. https://chat.openai.com
}
\author{
Lukas Soenning
}
