% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_brown.R
\docType{data}
\name{metadata_brown}
\alias{metadata_brown}
\title{Text metadata for Brown corpora}
\format{
\subsection{\code{metadata_brown}}{

A data frame with 500 rows and 3 columns:

\describe{
\item{text_file}{Standardized name of the text file (e.g. "A01", "J58", "R07")}
\item{macro_genre}{4 macro genres ("press", "general_prose", "learned", "fiction")}
\item{genre}{15 genres (e.g. "press_editorial", "popular_lore", "adventure_western_fiction"))}
}
}
}
\source{
McEnery, Tony & Andrew Hardie. 2012. \emph{Corpus linguistics}. Cambridge: Cambridge University Press.
}
\usage{
metadata_brown
}
\description{
This dataset provides metadata for the text files in the Brown family of corpora. It maps standardized file names to the textual categories macro genre and genre. Macro genres and genres are ordered based on the sampling frame informing the design of the Brown family of corpora (see https://listings.lib.msu.edu/public-corpora/cd421/manuals/brown/INDEX.HTM).
}
\keyword{datasets}
