% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_folded_power.R
\name{scale_x_fpower}
\alias{scale_x_fpower}
\title{Position scales for Tukey's folded power transformation}
\usage{
scale_x_fpower(lambda = 1, ...)
}
\arguments{
\item{lambda}{Numeric value of the applied power transformation}

\item{...}{Other argument passed on to \verb{scale_(x|y)_continuous()}}
}
\value{
The ggplot2 function \verb{scale_(x|y)_continuous()} with the appropriate transformation
}
\description{
Position scales for Tukey's folded power transformation
}
\examples{
if (require("ggplot2")) {
  ggplot(
    data = data.frame(
      dispersion = seq(0, 1, .01)),
    aes(x = dispersion)) +
    geom_dotplot() +
    scale_x_fpower(lambda = .5)
}
}
